rename sync_shard.rs to sync_shard_messages.rs
